Split extractor tests by concern
Address the remaining findings that still applied in weaver-cards.

- add rustdoc to import, parameter, and structure helpers
- remove the redundant Rust import pub-prefix trim
- replace the oversized extractor.rs test module with focused submodules
  and a shared common test harness

Validated with cargo test -p weaver-cards, cargo clippy -p
weaver-cards --all-targets --all-features -- -D warnings,
make check-fmt, make lint, and make test.
