--- crates/weaver-cards/src/tests/extractor.rs
@@ -1,485 +1,0 @@
-//! Unit tests for Tree-sitter-backed card extraction.
-
-use std::path::Path;
-
-use rstest::rstest;
-
-use crate::{
-    CardExtractionError, CardExtractionInput, CardSymbolKind, DetailLevel, GetCardResponse,
-    TreeSitterCardExtractor,
-};
-
-#[derive(Clone, Copy)]
-struct ExtractRequest<'a> {
-    path: &'a Path,
-    source: &'a str,
-    line: u32,
-    column: u32,
-    detail: DetailLevel,
-}
-
-#[derive(Clone, Copy)]
-struct SymbolExpectation<'a> {
-    request: ExtractRequest<'a>,
-    expected_kind: CardSymbolKind,
-    expected_name: &'a str,
-    expected_container: Option<&'a str>,
-}
-
-#[derive(Clone, Copy)]
-struct CaseSpec {
-    path: &'static Path,
-    source: &'static str,
-    line: u32,
-    column: u32,
-    kind: CardSymbolKind,
-    name: &'static str,
-    container: Option<&'static str>,
-}
-
-impl From<CaseSpec> for SymbolExpectation<'static> {
-    fn from(s: CaseSpec) -> Self {
-        SymbolExpectation {
-            request: ExtractRequest {
-                path: s.path,
-                source: s.source,
-                line: s.line,
-                column: s.column,
-                detail: DetailLevel::Structure,
-            },
-            expected_kind: s.kind,
-            expected_name: s.name,
-            expected_container: s.container,
-        }
-    }
-}
-
-impl<'a> From<ExtractRequest<'a>> for CardExtractionInput<'a> {
-    fn from(r: ExtractRequest<'a>) -> Self {
-        CardExtractionInput {
-            path: r.path,
-            source: r.source,
-            line: r.line,
-            column: r.column,
-            detail: r.detail,
-        }
-    }
-}
-
-fn run_extraction(request: ExtractRequest<'_>) -> Result<crate::SymbolCard, CardExtractionError> {
-    let path = super::absolute_test_path(request.path);
-    TreeSitterCardExtractor::new().extract(
-        ExtractRequest {
-            path: &path,
-            ..request
-        }
-        .into(),
-    )
-}
-
-fn extract(request: ExtractRequest<'_>) -> crate::SymbolCard {
-    run_extraction(request).expect("card extraction should succeed")
-}
-
-fn extract_error(request: ExtractRequest<'_>) -> CardExtractionError {
-    run_extraction(request).expect_err("card extraction should fail")
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ExpectedError {
-    UnsupportedLanguage,
-    PositionOutOfRange,
-    NoSymbolAtPosition,
-}
-
-fn error_matches(err: &CardExtractionError, expected: ExpectedError) -> bool {
-    matches!(
-        (err, expected),
-        (
-            CardExtractionError::UnsupportedLanguage { .. },
-            ExpectedError::UnsupportedLanguage
-        ) | (
-            CardExtractionError::PositionOutOfRange { .. },
-            ExpectedError::PositionOutOfRange
-        ) | (
-            CardExtractionError::NoSymbolAtPosition { .. },
-            ExpectedError::NoSymbolAtPosition
-        )
-    )
-}
-
-#[rstest]
-#[case(CaseSpec { path: Path::new("fixture.rs"), source: "/// Greets callers.\nfn greet(name: &str) -> usize {\n    let count = name.len();\n    count\n}\n", line: 2, column: 4, kind: CardSymbolKind::Function, name: "greet", container: None }.into())]
-#[case(CaseSpec { path: Path::new("fixture.rs"), source: "struct Widget {\n    name: String,\n}\n", line: 1, column: 8, kind: CardSymbolKind::Type, name: "Widget", container: None }.into())]
-#[case(CaseSpec { path: Path::new("fixture.rs"), source: "impl Widget {\n    fn render(&self) {}\n}\n", line: 2, column: 8, kind: CardSymbolKind::Method, name: "render", container: Some("Widget") }.into())]
-#[case(CaseSpec { path: Path::new("fixture.py"), source: "def greet(name: str) -> int:\n    total = len(name)\n    return total\n", line: 1, column: 5, kind: CardSymbolKind::Function, name: "greet", container: None }.into())]
-#[case(CaseSpec { path: Path::new("fixture.py"), source: "class Widget:\n    pass\n", line: 1, column: 7, kind: CardSymbolKind::Class, name: "Widget", container: None }.into())]
-#[case(CaseSpec { path: Path::new("fixture.py"), source: "class Widget:\n    def render(self) -> None:\n        status = True\n        if status:\n            return None\n", line: 2, column: 9, kind: CardSymbolKind::Method, name: "render", container: Some("Widget") }.into())]
-#[case(CaseSpec { path: Path::new("fixture.ts"), source: "function greet(name: string): number {\n  const total = name.length;\n  return total;\n}\n", line: 1, column: 10, kind: CardSymbolKind::Function, name: "greet", container: None }.into())]
-#[case(CaseSpec { path: Path::new("fixture.ts"), source: "interface Widget {\n  name: string;\n}\n", line: 1, column: 11, kind: CardSymbolKind::Interface, name: "Widget", container: None }.into())]
-#[case(CaseSpec { path: Path::new("fixture.ts"), source: "class Widget {\n  render(): void {\n    const ready = true;\n    if (ready) {\n      return;\n    }\n  }\n}\n", line: 2, column: 3, kind: CardSymbolKind::Method, name: "render", container: Some("Widget") }.into())]
-fn extracts_supported_symbol_kinds(#[case] case: SymbolExpectation<'static>) {
-    let card = extract(case.request);
-    assert_eq!(card.symbol.symbol_ref.kind, case.expected_kind);
-    assert_eq!(card.symbol.symbol_ref.name, case.expected_name);
-    assert_eq!(
-        card.symbol.symbol_ref.container.as_deref(),
-        case.expected_container
-    );
-}
-
-#[test]
-fn returns_module_cards_for_import_interstitials() {
-    let requests = [
-        ExtractRequest {
-            path: Path::new("fixture.rs"),
-            source: "use std::fmt;\nuse std::io;\n\nfn greet() {}\n",
-            line: 1,
-            column: 1,
-            detail: DetailLevel::Structure,
-        },
-        ExtractRequest {
-            path: Path::new("fixture.py"),
-            source: "import os\nfrom pkg import value\n\ndef greet() -> None:\n    pass\n",
-            line: 1,
-            column: 1,
-            detail: DetailLevel::Structure,
-        },
-        ExtractRequest {
-            path: Path::new("fixture.ts"),
-            source: "import { join } from 'node:path';\nimport type { Widget } from './widget';\n\nfunction greet(): void {}\n",
-            line: 1,
-            column: 1,
-            detail: DetailLevel::Structure,
-        },
-    ];
-
-    for request in requests {
-        let card = extract(request);
-        assert_eq!(card.symbol.symbol_ref.kind, CardSymbolKind::Module);
-        let interstitial = card
-            .interstitial
-            .as_ref()
-            .expect("module card should carry interstitial imports");
-        assert!(
-            !interstitial.imports.normalized.is_empty(),
-            "expected normalized imports, got {:?}",
-            interstitial.imports.normalized
-        );
-    }
-}
-
-#[test]
-fn extraction_ranges_are_deterministic() {
-    let request = ExtractRequest {
-        path: Path::new("fixture.rs"),
-        source: "fn greet(name: &str) -> usize {\n    name.len()\n}\n",
-        line: 1,
-        column: 4,
-        detail: DetailLevel::Structure,
-    };
-
-    let first = extract(request);
-    let second = extract(request);
-
-    assert_eq!(
-        first.symbol.symbol_ref.range,
-        second.symbol.symbol_ref.range
-    );
-    assert_eq!(first.etag, second.etag);
-}
-
-#[test]
-fn rust_comment_bundling_is_stable_under_whitespace_edits() {
-    let baseline = extract(ExtractRequest {
-        path: Path::new("fixture.rs"),
-        source: "/// Greets callers.\n/// Returns a count.\nfn greet(name: &str) -> usize {\n    name.len()\n}\n",
-        line: 3,
-        column: 4,
-        detail: DetailLevel::Structure,
-    });
-    let edited = extract(ExtractRequest {
-        path: Path::new("fixture.rs"),
-        source: "/// Greets callers.   \n/// Returns a count.\nfn greet(name: &str) -> usize {\n    name.len()\n}\n",
-        line: 3,
-        column: 4,
-        detail: DetailLevel::Structure,
-    });
-
-    assert_eq!(baseline.attachments, edited.attachments);
-    assert_eq!(baseline.doc, edited.doc);
-}
-
-#[test]
-fn decorator_bundling_is_stable_under_whitespace_edits() {
-    let baseline = extract(ExtractRequest {
-        path: Path::new("fixture.ts"),
-        source: "@sealed\nclass Widget {\n  render(): void {}\n}\n",
-        line: 2,
-        column: 7,
-        detail: DetailLevel::Structure,
-    });
-    let edited = extract(ExtractRequest {
-        path: Path::new("fixture.ts"),
-        source: "   @sealed   \nclass Widget {\n  render(): void {}\n}\n",
-        line: 2,
-        column: 7,
-        detail: DetailLevel::Structure,
-    });
-
-    assert_eq!(baseline.attachments, edited.attachments);
-}
-
-#[test]
-fn nested_locals_do_not_become_entities() {
-    let requests = [
-        ExtractRequest {
-            path: Path::new("fixture.rs"),
-            source: "fn outer() {\n    fn inner() {}\n    inner();\n}\n",
-            line: 2,
-            column: 8,
-            detail: DetailLevel::Structure,
-        },
-        ExtractRequest {
-            path: Path::new("fixture.py"),
-            source: "def outer() -> None:\n    def inner() -> None:\n        return None\n    inner()\n",
-            line: 2,
-            column: 9,
-            detail: DetailLevel::Structure,
-        },
-        ExtractRequest {
-            path: Path::new("fixture.ts"),
-            source: "function outer(): void {\n  function inner(): void {}\n  inner();\n}\n",
-            line: 2,
-            column: 12,
-            detail: DetailLevel::Structure,
-        },
-    ];
-
-    for request in requests {
-        let card = extract(request);
-        assert_eq!(card.symbol.symbol_ref.kind, CardSymbolKind::Function);
-        assert_eq!(card.symbol.symbol_ref.name, "outer");
-    }
-}
-
-#[test]
-fn whitespace_only_edits_do_not_change_symbol_id() {
-    let compact = extract(ExtractRequest {
-        path: Path::new("fixture.py"),
-        source: "def greet(name: str) -> int:\n    total = len(name)\n    return total\n",
-        line: 1,
-        column: 5,
-        detail: DetailLevel::Structure,
-    });
-    let spaced = extract(ExtractRequest {
-        path: Path::new("fixture.py"),
-        source: "def greet( name: str ) -> int:\n\n    total = len(name)\n    return total\n",
-        line: 1,
-        column: 5,
-        detail: DetailLevel::Structure,
-    });
-
-    assert_eq!(compact.symbol.symbol_id, spaced.symbol.symbol_id);
-}
-
-#[test]
-fn rust_tuple_type_parameters_are_extracted_from_ast() {
-    let card = extract(ExtractRequest {
-        path: Path::new("fixture.rs"),
-        source: "fn foo(pair: (u32, u32)) -> u32 {\n    pair.0 + pair.1\n}\n",
-        line: 1,
-        column: 4,
-        detail: DetailLevel::Signature,
-    });
-    let repeated = extract(ExtractRequest {
-        path: Path::new("fixture.rs"),
-        source: "fn foo(pair: (u32, u32)) -> u32 {\n    pair.0 + pair.1\n}\n",
-        line: 1,
-        column: 4,
-        detail: DetailLevel::Signature,
-    });
-    let params = &card.signature.as_ref().expect("signature").params;
-
-    assert_eq!(params.len(), 1);
-    assert_eq!(
-        params.first().map(|param| param.name.as_str()),
-        Some("pair")
-    );
-    assert_eq!(
-        params.first().map(|param| param.type_annotation.as_str()),
-        Some("(u32, u32)")
-    );
-    assert_eq!(card.symbol.symbol_id, repeated.symbol.symbol_id);
-}
-
-#[test]
-fn python_default_parameters_are_extracted_from_ast() {
-    let card = extract(ExtractRequest {
-        path: Path::new("fixture.py"),
-        source: "def bar(x, y=1):\n    return x + y\n",
-        line: 1,
-        column: 5,
-        detail: DetailLevel::Signature,
-    });
-    let repeated = extract(ExtractRequest {
-        path: Path::new("fixture.py"),
-        source: "def bar(x, y=1):\n    return x + y\n",
-        line: 1,
-        column: 5,
-        detail: DetailLevel::Signature,
-    });
-    let params = &card.signature.as_ref().expect("signature").params;
-
-    assert_eq!(params.len(), 2);
-    assert_eq!(params.first().map(|param| param.name.as_str()), Some("x"));
-    assert_eq!(params.get(1).map(|param| param.name.as_str()), Some("y"));
-    assert_eq!(
-        params
-            .iter()
-            .map(|param| param.type_annotation.as_str())
-            .collect::<Vec<_>>(),
-        vec!["", ""]
-    );
-    assert_eq!(card.symbol.symbol_id, repeated.symbol.symbol_id);
-}
-
-#[test]
-fn python_raw_triple_quoted_docstrings_are_preserved() {
-    let card = extract(ExtractRequest {
-        path: Path::new("fixture.py"),
-        source: "def bar() -> None:\n    r\"\"\"raw docstring\"\"\"\n    return None\n",
-        line: 1,
-        column: 5,
-        detail: DetailLevel::Structure,
-    });
-
-    assert_eq!(
-        card.doc.as_ref().map(|doc| doc.docstring.as_str()),
-        Some("raw docstring")
-    );
-}
-
-#[test]
-fn signature_display_preserves_literal_whitespace() {
-    let card = extract(ExtractRequest {
-        path: Path::new("fixture.ts"),
-        source: "function greet(message = \"a  b\"): void {\n  return;\n}\n",
-        line: 1,
-        column: 10,
-        detail: DetailLevel::Signature,
-    });
-
-    assert_eq!(
-        card.signature
-            .as_ref()
-            .map(|signature| signature.display.as_str()),
-        Some("function greet(message = \"a  b\"): void")
-    );
-}
-
-#[test]
-fn decorator_text_preserves_literal_whitespace() {
-    let card = extract(ExtractRequest {
-        path: Path::new("fixture.py"),
-        source: "@route(\"a  b\")\ndef bar() -> None:\n    return None\n",
-        line: 2,
-        column: 5,
-        detail: DetailLevel::Structure,
-    });
-
-    assert_eq!(
-        card.attachments
-            .as_ref()
-            .and_then(|attachments| attachments.decorators.first().map(String::as_str)),
-        Some("@route(\"a  b\")")
-    );
-}
-
-#[test]
-fn semantic_detail_degrades_to_tree_sitter_provenance() {
-    let card = extract(ExtractRequest {
-        path: Path::new("fixture.ts"),
-        source: "function greet(name: string): number {\n  return name.length;\n}\n",
-        line: 1,
-        column: 10,
-        detail: DetailLevel::Semantic,
-    });
-
-    assert!(card.lsp.is_none());
-    assert_eq!(
-        card.provenance.sources,
-        vec![
-            String::from("tree_sitter"),
-            String::from("tree_sitter_degraded_semantic"),
-        ]
-    );
-}
-
-#[test]
-fn get_card_success_payload_can_wrap_extracted_cards() {
-    let card = extract(ExtractRequest {
-        path: Path::new("fixture.rs"),
-        source: "fn greet() {}\n",
-        line: 1,
-        column: 4,
-        detail: DetailLevel::Minimal,
-    });
-    let response = GetCardResponse::Success {
-        card: Box::new(card),
-    };
-
-    assert!(matches!(response, GetCardResponse::Success { .. }));
-}
-
-#[rstest]
-#[case(
-    ExtractRequest { path: Path::new("fixture.foobar"), source: "fn main() {}\n", line: 1, column: 1, detail: DetailLevel::Full },
-    ExpectedError::UnsupportedLanguage,
-)]
-#[case(
-    ExtractRequest { path: Path::new("fixture.rs"), source: "fn main() {}\n", line: 0, column: 1, detail: DetailLevel::Full },
-    ExpectedError::PositionOutOfRange,
-)]
-#[case(
-    ExtractRequest { path: Path::new("fixture.rs"), source: "fn main() {}\n", line: 10, column: 100, detail: DetailLevel::Full },
-    ExpectedError::PositionOutOfRange,
-)]
-#[case(
-    ExtractRequest { path: Path::new("fixture.rs"), source: "// heading\nfn visible_symbol() {}\n", line: 1, column: 1, detail: DetailLevel::Full },
-    ExpectedError::NoSymbolAtPosition,
-)]
-fn extraction_error_cases(
-    #[case] request: ExtractRequest<'static>,
-    #[case] expected: ExpectedError,
-) {
-    let err = extract_error(request);
-
-    assert!(
-        error_matches(&err, expected),
-        "expected {expected:?}, got {err:?}",
-    );
-}
-
-#[test]
-fn returns_parse_error_when_parser_setup_fails() {
-    let path = super::absolute_test_path(Path::new("fixture.rs"));
-    let err = TreeSitterCardExtractor::extract_with_parser_for_test(
-        CardExtractionInput {
-            path: &path,
-            source: "fn main() {}\n",
-            line: 1,
-            column: 1,
-            detail: DetailLevel::Full,
-        },
-        |language| {
-            Err(CardExtractionError::Parse {
-                language: String::from(language.as_str()),
-                message: String::from("forced parse failure"),
-            })
-        },
-    )
-    .expect_err("expected parse error");
-
-    assert!(matches!(err, CardExtractionError::Parse { .. }));
-}
