+++ crates/weaver-cards/src/tests/extractor/mod.rs
@@ -1,0 +1,8 @@
+//! Unit tests for Tree-sitter-backed card extraction.
+
+mod common;
+mod error_cases_tests;
+mod general_tests;
+mod signature_decorator_tests;
+mod symbol_kind_tests;
+mod whitespace_stability_tests;
